refactor: Move lexer in its own module
