+++ src/core/mod.rs
@@ -1,5 +1,6 @@
 pub mod ast;
 pub mod constraints;
 pub mod model;
+pub mod lexer;
 pub mod parser;
 pub mod utils;
